feat: export tenant worker runtime and MCP server glue from packages
Tenant deployment repos (forks of apps/full-app) were byte-for-byte copying
two chunks of server code that should be importable from the published
packages. Relocate both so tenants import + configure instead of copy:

Worker runtime -> @hachej/boring-agent/server/worker
- Move worker auth/exec/routes/workspace + createWorkerServer factory into
  packages/agent under a new ./server/worker subpath export.
- Env reads centralized in src/server/config/workerConfig.ts and wire-protocol
  error codes moved to a WORKER_ERROR_CODES enum to satisfy agent invariants.
- full-app/agent-worker.ts is now a thin entry over createWorkerServer(); the
  local worker/ copies are deleted and the Dockerfile no longer copies them.

MCP server glue -> @hachej/boring-mcp/server
- Extract the source registry, HTTP error mapping, route registration, agent
  tools, and Composio transport into appServerBinding.ts as config-driven
  factories (connector configs, secret env vars, redaction canaries, client
  identity, and a whenDisabled skip/serve-503 policy are all parameters).
- boring-mcp now depends on @hachej/boring-core; full-app build order updated
  so core builds before mcp.
- full-app/boringMcp.ts keeps its public names as thin wrappers that supply
  only full-app's deployment config, proving the exports are sufficient.

Add boot tests for both new factories; existing full-app boringMcp tests pass
unchanged through the wrappers.
