+++ packages/agent/src/front/__tests__/publicExports.test.ts
@@ -30,6 +30,7 @@ describe('@hachej/boring-agent/front public exports', () => {
       './front',
       './front/styles.css',
       './server',
+      './server/worker',
       './shared',
     ])
   })
