+++ plugins/boring-mcp/package.json
@@ -53,12 +53,16 @@
     "clean": "rm -rf dist .tsbuildinfo"
   },
   "peerDependencies": {
+    "@hachej/boring-core": "workspace:*",
     "@hachej/boring-workspace": "workspace:*",
+    "fastify": "^5.9.0",
     "react": "^18.0.0 || ^19.0.0",
     "react-dom": "^18.0.0 || ^19.0.0"
   },
   "devDependencies": {
+    "@hachej/boring-core": "workspace:*",
     "@hachej/boring-workspace": "workspace:*",
+    "fastify": "^5.9.0",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.0",
     "@types/node": "^22.20.0",
